+++ internal/newsclient/list/list_test.go
@@ -1,4 +1,4 @@
-package everything
+package list
 
 import (
 	"context"
